hover-card test,story and props
