+++ packages/components/hover-card/src/index.ts
@@ -2,56 +2,25 @@ import type { } from '@floating-ui/vue'
 
 export {
   OkuHoverCard,
-  hoverCardProps,
 } from './hoverCard'
 
 export type {
-  HoverCardEmits,
-  HoverCardProps,
   /* HoverCardElement,
   HoverCardIntrinsicElement, */
 } from './hoverCard'
 
 export {
   OkuHoverCardTrigger,
-  hoverCardTriggerProps,
-} from './hoverCardTrigger'
-
-export type {
-  HoverCardTriggerElement,
-  HoverCardTriggerNativeElement,
-  HoverCardTriggerProps,
 } from './hoverCardTrigger'
 
 export {
   OkuHoverCardPortal,
-  hoverCardPortalProps,
-} from './hoverCardPortal'
-
-export type {
-  HoverCardPortalProps,
-  HoverCardPortalElement,
 } from './hoverCardPortal'
 
 export {
   OkuHoverCardContent,
-  hoverCardContentProps,
-} from './hoverCardContent'
-
-export type {
-  HoverCardContentElement,
-  HoverCardContentNaviteElement,
-  HoverCardContentProps,
-  HoverCardContentEmits,
 } from './hoverCardContent'
 
 export {
   OkuHoverCardArrow,
-  hoverCardArrowProps,
-} from './hoverCardArrow'
-
-export type {
-  HoverCardArrowElement,
-  HoverCardArrowNaviteElement,
-  HoverCardArrowProps,
 } from './hoverCardArrow'
